requirements.txt
discord.py
python-dotenv
psutil
openai
pytest
pytest-asyncio
pytest-cov
